Cambios que no me acuerdo que eran jeje

+++ Progra1BD/Models/Account.cs
@@ -4,11 +4,11 @@ namespace Progra1BD.Models
     {
         public int ID { get; set; }
         
-        public int idCliente { get; set; }
+        public long idCliente { get; set; }
         
         public int idTipoCuenta { get; set; }
         
-        public int numCuenta { get; set; }
+        public long numCuenta { get; set; }
         
         public float saldo { get; set; }
     }
